refactor: Separate LogonForm to own component
